+++ tests/testthat/test-flair.R
@@ -1,31 +1,31 @@
 test_str = "ggplot(iris, aes(x = Sepal.Length)) + geom_histogram()"
 test_regexp = "Sepal\\.[:alnum:]*"
 
-test_that("hlt_regexp works without dots", {
+test_that("flair_rx works without dots", {
 
   good_str = "ggplot(iris, aes(x = <span style='background-color:#ffff7f'>Sepal.Length</span>)) + geom_histogram()"
 
-  expect_equal(hlt_regexp(test_str, test_regexp), good_str)
+  expect_equal(flair_rx(test_str, test_regexp), good_str)
 })
 
 
 
-test_that("hlt_regexp works with dots", {
+test_that("flair_rx works with dots", {
 
   good_str = "ggplot(iris, aes(x = <span style='color:red;font-size:30px'>Sepal.Length</span>)) + geom_histogram()"
 
-  expect_equal(hlt_regexp(test_str, test_regexp, color = "red", size = "30px"), good_str)
+  expect_equal(flair_rx(test_str, test_regexp, color = "red", size = "30px"), good_str)
 })
 
 
-test_that("hlt_regexp works for flair code", {
+test_that("flair_rx works for decorated code", {
 
   good_str = "ggplot(iris, aes(x = <span style='color:red;font-size:30px'>Sepal.Length</span>)) + geom_histogram()"
 
-  test_dc <- flair_code('ggplot(iris, aes(x = Sepal.Length)) + geom_histogram()')
+  test_dc <- decorate('ggplot(iris, aes(x = Sepal.Length)) + geom_histogram()')
 
-  test_result <- hlt_regexp(test_dc, test_regexp, color = "red", size = "30px")
+  test_result <- flair_rx(test_dc, test_regexp, color = "red", size = "30px")
 
   expect_equal(test_result[[1]], good_str)
-  expect_equal(class(test_result), "flair_code")
+  expect_equal(class(test_result), "decorate_code")
 })